refactor: minimize PR diff to extraction-only changes
Restore the non-extracted CLI code to its original main location and content,
so the diff reflects only code extracted into @deepnote/database-integrations
(plus CLI imports updated to consume it), not incidental code movement.

- utils/dotenv.ts: restored verbatim from main (not extracted)
- integrations/inject-integration-env-vars.ts: back to main content
- utils/env-var-refs.ts + .test.ts: keep only resolveEnvVarRefs in place,
  importing the extracted resolveEnvVarRefsFromMap from the library
- readIntegrationsDocument/writeIntegrationsFile stay in commands/integrations.ts;
  drop the interim integrations-document.ts module
- run.test.ts / merge-integrations.test.ts: re-point only the extracted symbols
  to the library, keep CLI-local mocks/imports as on main
