+++ packages/cli/src/commands/integrations.ts
@@ -1,3 +1,5 @@
+import fs from 'node:fs/promises'
+import path from 'node:path'
 import {
   ApiError,
   createNewDocument,
@@ -7,16 +9,18 @@ import {
   DEFAULT_INTEGRATIONS_FILE,
   fetchIntegrations,
   mergeApiIntegrationsIntoDocument,
+  parseIntegrationsDocument,
   SCHEMA_COMMENT,
+  serializeIntegrationsDocument,
 } from '@deepnote/database-integrations'
 import chalk from 'chalk'
 import type { Command } from 'commander'
-import { isSeq } from 'yaml'
+import { type Document, isSeq } from 'yaml'
 import { ExitCode } from '../exit-codes'
-import { readIntegrationsDocument, writeIntegrationsFile } from '../integrations/integrations-document'
 import { debug, log, output } from '../output'
 import { MissingTokenError } from '../utils/auth'
 import { updateDotEnv } from '../utils/dotenv'
+import { isErrnoENOENT } from '../utils/file-resolver'
 
 // ============================================================================
 // Options Interface
@@ -43,6 +47,37 @@ function resolveToken(options: IntegrationsPullOptions): string {
     throw new MissingTokenError()
   }
   return token
+}
+
+/**
+ * Read existing integrations file as a YAML Document.
+ * Returns null if file doesn't exist or is empty.
+ * This preserves comments and formatting for later manipulation.
+ */
+export async function readIntegrationsDocument(filePath: string): Promise<Document | null> {
+  try {
+    const content = await fs.readFile(filePath, 'utf-8')
+    return parseIntegrationsDocument(content)
+  } catch (error) {
+    if (isErrnoENOENT(error)) {
+      return null
+    }
+    throw error
+  }
+}
+
+/**
+ * Write integrations document to the YAML file.
+ * Uses doc.toString() to preserve comments and formatting.
+ */
+export async function writeIntegrationsFile(filePath: string, doc: Document): Promise<void> {
+  const yamlContent = serializeIntegrationsDocument(doc)
+
+  // Ensure parent directory exists
+  const dir = path.dirname(filePath)
+  await fs.mkdir(dir, { recursive: true })
+
+  await fs.writeFile(filePath, yamlContent, 'utf-8')
 }
 
 /**
