+++ packages/cli/src/commands/integrations/add-integration.ts
@@ -14,9 +14,9 @@ import { input, select } from '@inquirer/prompts'
 import chalk from 'chalk'
 import type { Command } from 'commander'
 import { ExitCode } from '../../exit-codes'
-import { readIntegrationsDocument, writeIntegrationsFile } from '../../integrations/integrations-document'
 import { log, output } from '../../output'
 import { updateDotEnv } from '../../utils/dotenv'
+import { readIntegrationsDocument, writeIntegrationsFile } from '../integrations'
 import { promptForFieldsAlloydb } from './integrations-prompts/alloydb'
 import { promptForFieldsAthena } from './integrations-prompts/athena'
 import { promptForFieldsBigQuery } from './integrations-prompts/big-query'
