+++ packages/cli/src/commands/integrations/edit-integration.ts
@@ -16,10 +16,10 @@ import type { Command } from 'commander'
 import { type Document, isMap, isSeq, type YAMLMap } from 'yaml'
 import z from 'zod'
 import { ExitCode } from '../../exit-codes'
-import { readIntegrationsDocument, writeIntegrationsFile } from '../../integrations/integrations-document'
 import { debug, log, output } from '../../output'
 import { readDotEnv, updateDotEnv } from '../../utils/dotenv'
 import { getProcessEnv } from '../../utils/process-env'
+import { readIntegrationsDocument, writeIntegrationsFile } from '../integrations'
 import { promptForIntegrationName } from './add-integration'
 import { promptForFieldsAlloydb } from './integrations-prompts/alloydb'
 import { promptForFieldsAthena } from './integrations-prompts/athena'
