+++ packages/cli/src/integrations/inject-integration-env-vars.ts
@@ -1,8 +1,9 @@
 import { type DatabaseIntegrationConfig, getEnvironmentVariablesForIntegrations } from '@deepnote/database-integrations'
+import { debug } from '../output'
 
 /**
- * Generate environment variables for the given integrations and inject them into `process.env`.
- * Returns the list of injected env var names.
+ * Generate environment variables for the given integrations and inject them into process.env.
+ * Returns the list of injected env var names (useful for testing/debugging).
  */
 export function injectIntegrationEnvVars(
   integrations: DatabaseIntegrationConfig[],
@@ -12,13 +13,21 @@ export function injectIntegrationEnvVars(
     return []
   }
 
-  const { envVars } = getEnvironmentVariablesForIntegrations(integrations, {
+  const { envVars, errors } = getEnvironmentVariablesForIntegrations(integrations, {
     projectRootDirectory: workingDirectory,
   })
 
+  // Log any errors from env var generation
+  for (const error of errors) {
+    debug(`Integration env var error: ${error.message}`)
+  }
+
+  // Inject env vars into process.env
   for (const { name, value } of envVars) {
     process.env[name] = value
   }
+
+  debug(`Injected ${envVars.length} environment variables for integrations`)
 
   return envVars.map(v => v.name)
 }
