+++ packages/cli/src/integrations/merge-integrations.test.ts
@@ -10,7 +10,7 @@ import {
   mergeProcessedIntegrations,
 } from '@deepnote/database-integrations'
 import { afterAll, beforeAll, describe, expect, it } from 'vitest'
-import { readIntegrationsDocument, writeIntegrationsFile } from './integrations-document'
+import { readIntegrationsDocument, writeIntegrationsFile } from '../commands/integrations'
 
 // Helper to create a mock API integration
 function createMockApiIntegration(overrides: Partial<ApiIntegration> = {}): ApiIntegration {
